added function make_tags which creates tag dictionary from words, assigning colors and sizes
make_tags assigns sizes (based on count supplied with the words and minimum and maximum sizes)
and colors (completely random or randomly chosen from user-supplied list) to words.
The default scaling function is the square root - this is theoretically the correct scaling
if word count is mapped to tag area but can be changed.